+++ src/paint/PointBrush.java
@@ -5,6 +5,6 @@ public class PointBrush extends Brush
 	@Override
 	public void apply(Bitmap<Integer> b)
 	{
-		b.setPixel(x, y, color);
+		b.setPixel((int) x, (int) y, color);
 	}
 }
